+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 ## [Unreleased] - 
 ### Fixed
 - Allow mixture of unix and windows path slash in config file
+### Added
+- New output module
 
 ## [1.3.2] - 2022-10-20
 ### Added