+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 ### Added
 - New output module
 - New release module
+### Changed
+- Output module is now called at the end of each timestep
 
 ## [1.3.2] - 2022-10-20
 ### Added