+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2024-01-25
+### Fixed
+- Package now works with pandas 2.2.0
+
+
 ## [1.3.2] - 2022-10-20
 ### Added
 - Automatically publish to GitHub Releases and PyPI